refactore light and blur shadow
