+++ index.html
@@ -36,6 +36,8 @@
         width: 4096,
       }
       let model = null
+      let newGltfLoaded = false
+      let usingThreeShadow = false
       let modelTransform = {
         scale: 0.5,
         scaleX: 0.5,
